v1/requirements.txt
noise==1.2.2
pillow==9.2.0
numpy==1.23.1